Added `[output directory]` command line argument.
Now there is an additional required command line argument, for
specifying the directory to output results to. This required some
re-writing of `main.py`, as well as all of the individual "code" modules
and functions. The README has been edited to reflect these changes,
along with aesthetic changes.

There is a new `utils` module, which currently only contains the
function `make_sure_path_exists`. This is used to create the output
directories if they do not exist.

I have also taken the liberty to add doc-strings to all three of the
"code" modules, and cleaned up Jake's code a little bit.