Working on getting error bars on rate and fitted rate
There is likely something wrong with the way I'm using statsmodels'
weighted least squares (WLS) function. Using regular np.linalg.lstsq
gives a good fit, but with WLS either the slope or intercept is off. The
problem may be in the weights I'm using, or it may be something else.